+++ src/modals/RemoveParticipantFromCourseModal.tsx
@@ -2,9 +2,9 @@ import { useTranslation } from 'react-i18next';
 import { SubcourseParticipant } from '../types/lernfair/Course';
 import { getSchoolTypeKey } from '../types/lernfair/SchoolType';
 import { getGradeLabel } from '../Utility';
-import { type BaseModalProps, Modal, ModalFooter, ModalHeader, ModalTitle } from '@/components/atoms/Modal';
-import { Typography } from '@/components/atoms/Typography';
-import { Button } from '@/components/atoms/Button';
+import { type BaseModalProps, Modal, ModalFooter, ModalHeader, ModalTitle } from '@/components/Modal';
+import { Typography } from '@/components/Typography';
+import { Button } from '@/components/Button';
 import { gql } from '@/gql';
 import { useMutation } from '@apollo/client';
 import { toast } from 'sonner';
