+++ src/pages/student/single-course/StudentCourseButtons.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom';
 import { Lecture, Subcourse } from '../../../gql/graphql';
 import OpenCourseChatButton from '../../subcourse/OpenCourseChatButton';
 import VideoButton from '@/components/VideoButton';
-import { Button } from '@/components/atoms/Button';
+import { Button } from '@/components/Button';
 import { IconPencil } from '@tabler/icons-react';
 
 type ActionButtonProps = {
