+++ src/pages/subcourse/SubcourseData.tsx
@@ -7,8 +7,8 @@ import { TrafficStatus } from '../../types/lernfair/Course';
 import Utility, { getGradeLabel, getTrafficStatus } from '../../Utility';
 import AlertMessage from '../../widgets/AlertMessage';
 import CourseTrafficLamp from '../../widgets/CourseTrafficLamp';
-import { Typography } from '@/components/atoms/Typography';
-import { Badge } from '@/components/atoms/Badge';
+import { Typography } from '@/components/Typography';
+import { Badge } from '@/components/Badge';
 
 type SubcourseDataProps = {
     course: Pick<Course, 'name' | 'image'> & { shared?: boolean; tags: Pick<Course_Tag, 'name'>[] };
