+++ src/widgets/PromoteBanner.tsx
@@ -2,9 +2,9 @@ import CourseTrafficLamp from './CourseTrafficLamp';
 import CheckIcon from '../assets/icons/lernfair/Icon_Done.svg';
 import CallIcon from '../assets/icons/lernfair/Icon_Call.svg';
 import { useTranslation } from 'react-i18next';
-import { Separator } from '@/components/atoms/Separator';
-import { Typography } from '@/components/atoms/Typography';
-import { Button } from '@/components/atoms/Button';
+import { Separator } from '@/components/Separator';
+import { Typography } from '@/components/Typography';
+import { Button } from '@/components/Button';
 import { useMutation } from '@apollo/client';
 import { gql } from '@/gql';
 import { Subcourse } from 'gql/graphql';
